requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
fonttools==4.61.1
fredapi==0.5.2
frozendict==2.4.7
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
multitasking==0.0.12
numpy==2.4.2
packaging==26.0
pandas==3.0.1
peewee==4.0.0
pillow==12.1.1
platformdirs==4.9.2
protobuf==7.34.0
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
yfinance==1.2.0
